clean Benchmark folder
remove unnecessary files
rename .cpp file to .cc file
